TV: Move PodcastDetailViewModel to its own file
